+++ takvimi-wearos-compose/src/main/java/com/bqubique/takvimi_wearos_compose/view/home/HomeScreen.kt
@@ -4,9 +4,6 @@ import android.app.AlarmManager
 import android.app.PendingIntent
 import android.util.Log
 import androidx.compose.foundation.layout.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Alarm
-import androidx.compose.material.icons.outlined.Cancel
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.livedata.observeAsState
@@ -21,14 +18,10 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.wear.compose.material.*
-import com.bqubique.takvimi_wearos_compose.util.getNextPrayer
-import com.bqubique.takvimi_wearos_compose.util.getPrayerName
-import com.bqubique.takvimi_wearos_compose.util.getRandomImage
-import com.bqubique.takvimi_wearos_compose.util.getTimeLeft
+import com.bqubique.takvimi_wearos_compose.util.*
 import com.bqubique.takvimi_wearos_compose.view.theme.md_theme_dark_primary
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
-import java.util.*
 
 const val TAG = "HomeScreen"
 
@@ -138,81 +131,4 @@ fun HomeScreen(
             CircularProgressIndicator()
         }
     }
-}
-
-fun setAlarms(
-    prayerTimes: List<String?>,
-    alarmManager: AlarmManager,
-    pendingIntentMap: Map<String, PendingIntent>,
-) {
-    val listOfCalendars = mutableListOf<Calendar>()
-
-    for (i in prayerTimes.indices) {
-        listOfCalendars.add(Calendar.getInstance().apply {
-            timeInMillis = System.currentTimeMillis()
-            set(Calendar.HOUR_OF_DAY, prayerTimes[i]?.substring(0, 2)?.toInt()!!)
-            set(Calendar.MINUTE, prayerTimes[i]?.substring(5)?.toInt()!!)
-            set(Calendar.SECOND, 0)
-        })
-    }
-
-    alarmManager.setExactAndAllowWhileIdle(
-        AlarmManager.RTC_WAKEUP, listOfCalendars[1].timeInMillis, pendingIntentMap["sunrise"]
-    )
-    alarmManager.setExactAndAllowWhileIdle(
-        AlarmManager.RTC_WAKEUP,
-        listOfCalendars[1].timeInMillis - 900000,
-        pendingIntentMap["preSunrise"]
-    )
-    alarmManager.setExactAndAllowWhileIdle(
-        AlarmManager.RTC_WAKEUP, listOfCalendars[2].timeInMillis, pendingIntentMap["dhuhr"]
-    )
-    alarmManager.setExactAndAllowWhileIdle(
-        AlarmManager.RTC_WAKEUP,
-        listOfCalendars[2].timeInMillis - 900000,
-        pendingIntentMap["preDhuhr"]
-    )
-    alarmManager.setExactAndAllowWhileIdle(
-        AlarmManager.RTC_WAKEUP, listOfCalendars[3].timeInMillis, pendingIntentMap["asr"]
-    )
-    alarmManager.setExactAndAllowWhileIdle(
-        AlarmManager.RTC_WAKEUP,
-        listOfCalendars[3].timeInMillis - 900000,
-        pendingIntentMap["preAsr"]
-    )
-    alarmManager.setExactAndAllowWhileIdle(
-        AlarmManager.RTC_WAKEUP, listOfCalendars[4].timeInMillis, pendingIntentMap["maghrib"]
-    )
-    alarmManager.setExactAndAllowWhileIdle(
-        AlarmManager.RTC_WAKEUP,
-        listOfCalendars[4].timeInMillis - 900000,
-        pendingIntentMap["preMaghrib"]
-    )
-    alarmManager.setExactAndAllowWhileIdle(
-        AlarmManager.RTC_WAKEUP, listOfCalendars[5].timeInMillis, pendingIntentMap["isha"]
-    )
-    alarmManager.setExactAndAllowWhileIdle(
-        AlarmManager.RTC_WAKEUP,
-        listOfCalendars[5].timeInMillis - 900000,
-        pendingIntentMap["preIsha"]
-    )
-
-    Log.i(TAG, "setAlarms: Intents set for all prayer times and pre-prayer times.")
-}
-
-fun cancelAlarms(
-    alarmManager: AlarmManager,
-    pendingIntentMap: Map<String, PendingIntent>,
-) {
-    alarmManager.cancel(pendingIntentMap["sunrise"])
-    alarmManager.cancel(pendingIntentMap["preSunrise"])
-    alarmManager.cancel(pendingIntentMap["dhuhr"])
-    alarmManager.cancel(pendingIntentMap["preDhuhr"])
-    alarmManager.cancel(pendingIntentMap["asr"])
-    alarmManager.cancel(pendingIntentMap["preAsr"])
-    alarmManager.cancel(pendingIntentMap["maghrib"])
-    alarmManager.cancel(pendingIntentMap["preMaghrib"])
-    alarmManager.cancel(pendingIntentMap["isha"])
-    alarmManager.cancel(pendingIntentMap["preIsha"])
-    Log.i(TAG, "setAlarms: All intents canceled.")
 }
